[🚚Rename]: useTabStore import 경로 변경
index폴더를 만들어서 그 안에다가 store들을 관리하고 있습니다
