+++ src/components/organisms/LoginInput/LoginInput.jsx
@@ -1,6 +1,6 @@
 import LoginButton from '@/components/atoms/LoginButton/LoginButton';
 import { InputField } from '@/components/molecules';
-import { useTabStore } from '@/store/useTabStore';
+import { useTabStore } from '@/store';
 import pb from '@/utils/pocketbase';
 import { useRef } from 'react';
 import { useState } from 'react';
